+++ packages/preact_signals/lib/src/computed.dart
@@ -9,7 +9,7 @@ import 'readonly.dart';
 ///
 /// The returned computed signal is read-only, and its value is automatically
 /// updated when any signals accessed from within the callback function change.
-class Computed<T> extends ReadonlySignal<T> implements Listenable {
+class Computed<T> with Listenable, ReadonlySignal<T> {
   @internal
   T Function() fn;
 
@@ -74,14 +74,14 @@ class Computed<T> extends ReadonlySignal<T> implements Listenable {
     // Mark this computed signal running before checking the dependencies for value
     // changes, so that the RUNNING flag can be used to notice cyclical dependencies.
     flags |= RUNNING;
-    if (version > 0 && !needsToRecompute(this)) {
+    if (version > 0 && !needsToRecompute()) {
       flags &= ~RUNNING;
       return true;
     }
 
     final prevContext = evalContext;
     try {
-      prepareSources(this);
+      prepareSources();
       evalContext = this;
       final val = fn();
       if (!_isInitialized ||
@@ -98,7 +98,7 @@ class Computed<T> extends ReadonlySignal<T> implements Listenable {
       version++;
     }
     evalContext = prevContext;
-    cleanupSources(this);
+    cleanupSources();
     flags &= ~RUNNING;
     return true;
   }
@@ -114,14 +114,14 @@ class Computed<T> extends ReadonlySignal<T> implements Listenable {
         node.source.subscribeToNode(node);
       }
     }
-    ReadonlySignal.internalSubscribe(this, node);
+    internalSubscribe(node);
   }
 
   @override
   void unsubscribeFromNode(Node node) {
     // Only run the unsubscribe step if the computed signal has any subscribers.
     if (targets != null) {
-      signalUnsubscribe(this, node);
+      signalUnsubscribe(node);
 
       // Computed signal unsubscribes from its dependencies when it loses its last subscriber.
       // This makes it possible for unreferences subgraphs of computed signals to get garbage collected.
@@ -152,7 +152,7 @@ class Computed<T> extends ReadonlySignal<T> implements Listenable {
       throw Exception('Cycle detected');
     }
 
-    final node = addDependency(this);
+    final node = addDependency();
     internalRefresh();
     if (node != null) {
       node.version = version;
@@ -168,7 +168,7 @@ class Computed<T> extends ReadonlySignal<T> implements Listenable {
 
   @override
   void Function() subscribe(void Function(T value) fn) {
-    return signalSubscribe(this, fn);
+    return signalSubscribe(fn);
   }
 }