+++ packages/preact_signals/lib/src/signal.dart
@@ -6,7 +6,7 @@ import 'node.dart';
 import 'readonly.dart';
 
 /// Instance of a new plain signal
-class Signal<T> extends ReadonlySignal<T> {
+class Signal<T> with ReadonlySignal<T> {
   @override
   final int globalId;
 
@@ -52,22 +52,22 @@ class Signal<T> extends ReadonlySignal<T> {
 
   @override
   void subscribeToNode(Node node) {
-    ReadonlySignal.internalSubscribe(this, node);
+    internalSubscribe(node);
   }
 
   @override
   void unsubscribeFromNode(Node node) {
-    signalUnsubscribe(this, node);
+    signalUnsubscribe(node);
   }
 
   @override
   void Function() subscribe(void Function(T value) fn) {
-    return signalSubscribe(this, fn);
+    return signalSubscribe(fn);
   }
 
   @override
   T get value {
-    final node = addDependency(this);
+    final node = addDependency();
     if (node != null) {
       node.version = version;
     }