+++ packages/preact_signals/lib/src/signals.dart
@@ -1,9 +1,0 @@
-// part 'batch.dart';
-// part 'globals.dart';
-// part 'node.dart';
-// part 'untracked.dart';
-// part 'readonly.dart';
-// part 'signal.dart';
-// part 'computed.dart';
-// part 'effect.dart';
-// part 'listenable.dart';